+++ RELEASES.md
@@ -1,3 +1,15 @@
-## Astra Monitor 1 - January 10 2024
+# Astra Monitor 2 - January 12 2024
+
+### New features
+- Added an option to change Theme Style (Light/Dark) to better fit your desktop environment's theme
+
+### Bug fixes
+- Sensor header width not consistant [#1](https://github.com/AstraExt/astra-monitor/issues/1)
+- Sensor menu not centered
+- Sensor menu missing footer utility buttons [[#4](https://github.com/AstraExt/astra-monitor/issues/4)]
+- Icon madness fixed. Now icons are more unified and support light and dark themes
+- Icons missing on the preferences window [[#2](https://github.com/AstraExt/astra-monitor/issues/2)]
+
+# Astra Monitor 1 - January 10 2024
 
 Public release.